+++ src/main/java/me/desht/pneumaticcraft/client/gui/widget/ITooltipProvider.java
@@ -3,5 +3,5 @@ package me.desht.pneumaticcraft.client.gui.widget;
 import java.util.List;
 
 public interface ITooltipProvider {
-    void addTooltip(int mouseX, int mouseY, List<String> curTip, boolean shift);
+    void addTooltip(double mouseX, double mouseY, List<String> curTip, boolean shift);
 }
