+++ src/main/java/me/desht/pneumaticcraft/client/gui/widget/WidgetCheckBox.java
@@ -105,7 +105,7 @@ public class WidgetCheckBox extends Widget implements ITaggedWidget, ITooltipPro
     }
 
     @Override
-    public void addTooltip(int mouseX, int mouseY, List<String> curTip, boolean shift) {
+    public void addTooltip(double mouseX, double mouseY, List<String> curTip, boolean shift) {
         curTip.addAll(tooltip);
     }
 }
