+++ src/main/java/me/desht/pneumaticcraft/client/gui/widget/WidgetTemperature.java
@@ -52,7 +52,7 @@ public class WidgetTemperature extends Widget implements ITooltipProvider {
     }
 
     @Override
-    public void addTooltip(int mouseX, int mouseY, List<String> curTip, boolean shift) {
+    public void addTooltip(double mouseX, double mouseY, List<String> curTip, boolean shift) {
         curTip.add("Temperature: " + (logic.getTemperatureAsInt() - 273) + "\u00b0C");
     }
 }
